Moved command handler to root folder
